File Check_results.py in Regression/Postprocessing was added. It represents the checking procedure of performance on arousal-recognition task.
Sme files was slightly modified to make training process on colab stable.
One critical fix: original_sample_rate in Rgression/Training/train_raw_data.py was replaced from 5 to 10.

+++ Audio/utils/utils.py
@@ -104,7 +104,7 @@ def plot_confusion_matrix(y_true, y_pred, classes,
 
 def load_labels(path_to_labels):
     f = open(path_to_labels, 'r')
-    original_sample_rate = int(f.readline().split(':')[-1])
+    original_sample_rate = float(f.readline().split(':')[-1])
     f.close()
     labels=pd.read_csv(path_to_labels, skiprows=1,header=None)
     return labels.values, original_sample_rate
